implement zoomToParent within phylotree.changes
