+++ src/components/tree/phyloTree/layouts.js
@@ -1,5 +1,6 @@
 /* eslint-disable no-multi-spaces */
-import { min, sum } from "d3-array";
+/* eslint-disable space-infix-ops */
+import { min, max, sum } from "d3-array";
 import { addLeafCount } from "./helpers";
 import { timerStart, timerEnd } from "../../../util/perf";
 
@@ -256,4 +257,103 @@ export const setScales = function setScales(margins) {
       this.yScale.range([height - (margins["bottom"] || 0), margins["top"] || 0]);
     }
   }
+};
+
+
+/**
+* this function sets the xScale, yScale domains and maps precalculated x,y
+* coordinates to their places on the screen
+* @return {null}
+*/
+export const mapToScreen = function mapToScreen() {
+  console.log("mapToScreen")
+  timerStart("mapToScreen");
+  /* set the range of the x & y scales */
+  this.setScales(this.params.margins);
+
+  /* find minimum & maximum x & y values, as well as # tips in view */
+  this.nNodesInView = 0;
+  let [minY, maxY, minX, maxX] = [1000000, 0, 1000000, 0];
+  this.nodes.filter((d) => d.inView).forEach((d) => {
+    if (d.x > maxX) maxX = d.x;
+    if (d.y > maxY) maxY = d.y;
+    if (d.x < minX) minX = d.x;
+    if (d.y < minY) minY = d.y;
+    if (d.terminal) this.nNodesInView++;
+  });
+
+  /* set the domain of the x & y scales */
+  if (this.layout === "radial" || this.layout === "unrooted") {
+    // handle "radial and unrooted differently since they need to be square
+    // since branch length move in x and y direction
+    // TODO: should be tied to svg dimensions
+    const spanX = maxX-minX;
+    const spanY = maxY-minY;
+    const maxSpan = max([spanY, spanX]);
+    const ySlack = (spanX>spanY) ? (spanX-spanY)*0.5 : 0.0;
+    const xSlack = (spanX<spanY) ? (spanY-spanX)*0.5 : 0.0;
+    this.xScale.domain([minX-xSlack, minX+maxSpan-xSlack]);
+    this.yScale.domain([minY-ySlack, minY+maxSpan-ySlack]);
+  } else if (this.layout==="clock") {
+    // same as rectangular, but flipped yscale
+    this.xScale.domain([minX, maxX]);
+    this.yScale.domain([maxY, minY]);
+  } else { // rectangular
+    this.xScale.domain([minX, maxX]);
+    this.yScale.domain([minY, maxY]);
+  }
+
+  // pass all x,y through scales and assign to xTip, xBase
+  this.nodes.forEach((d) => {
+    d.xTip = this.xScale(d.x);
+    d.yTip = this.yScale(d.y);
+    d.xBase = this.xScale(d.px);
+    d.yBase = this.yScale(d.py);
+  });
+  if (this.vaccines) {
+    this.vaccines.forEach((d) => {
+      const n = 5; /* half the number of pixels that the cross will take up */
+      const xTipCross = this.xScale(d.xCross); /* x position of the center of the cross */
+      const yTipCross = this.yScale(d.yCross); /* x position of the center of the cross */
+      d.vaccineCross = ` M ${xTipCross-n},${yTipCross-n} L ${xTipCross+n},${yTipCross+n} M ${xTipCross-n},${yTipCross+n} L ${xTipCross+n},${yTipCross-n}`;
+      d.vaccineLine = ` M ${d.xTip},${d.yTip} L ${xTipCross},${yTipCross}`;
+    });
+  }
+  if (this.params.confidence && this.layout==="rect") {
+    this.nodes.forEach((d) => {d.xConf = [this.xScale(d.conf[0]), this.xScale(d.conf[1])];});
+  }
+
+  // assign the branches as path to each node for the different layouts
+  if (this.layout==="clock" || this.layout==="unrooted") {
+    this.nodes.forEach((d) => {
+      d.branch = [" M "+d.xBase.toString()+","+d.yBase.toString()+" L "+d.xTip.toString()+","+d.yTip.toString(), ""];
+    });
+  } else if (this.layout==="rect") {
+    this.nodes.forEach((d) => {
+      const stem_offset = 0.5*(d.parent["stroke-width"] - d["stroke-width"]) || 0.0;
+      const childrenY = [this.yScale(d.yRange[0]), this.yScale(d.yRange[1])];
+      d.branch =[` M ${d.xBase - stem_offset},${d.yBase} L ${d.xTip},${d.yTip} M ${d.xTip},${childrenY[0]} L ${d.xTip},${childrenY[1]}`];
+      if (this.params.confidence) d.confLine =` M ${d.xConf[0]},${d.yBase} L ${d.xConf[1]},${d.yTip}`;
+    });
+  } else if (this.layout==="radial") {
+    const offset = this.nodes[0].depth;
+    const stem_offset_radial = this.nodes.map((d) => {return (0.5*(d.parent["stroke-width"] - d["stroke-width"]) || 0.0);});
+    this.nodes.forEach((d) => {d.cBarStart = this.yScale(d.yRange[0]);});
+    this.nodes.forEach((d) => {d.cBarEnd = this.yScale(d.yRange[1]);});
+    this.nodes.forEach((d, i) => {
+      d.branch =[
+        " M "+(d.xBase-stem_offset_radial[i]*Math.sin(d.angle)).toString()
+        + " "+(d.yBase-stem_offset_radial[i]*Math.cos(d.angle)).toString()
+        + " L "+d.xTip.toString()+" "+d.yTip.toString(), ""
+      ];
+      if (!d.terminal) {
+        d.branch[1] =[" M "+this.xScale(d.xCBarStart).toString()+" "+this.yScale(d.yCBarStart).toString()+
+        " A "+(this.xScale(d.depth)-this.xScale(offset)).toString()+" "
+        +(this.yScale(d.depth)-this.yScale(offset)).toString()
+        +" 0 "+(d.smallBigArc?"1 ":"0 ") +" 1 "+
+        " "+this.xScale(d.xCBarEnd).toString()+","+this.yScale(d.yCBarEnd).toString()];
+      }
+    });
+  }
+  timerEnd("mapToScreen");
 };
